Extract AddParentBiometricsAttributeValueIfRequired to a separate file in kotlin
