+++ app/src/main/java/org/dhis2/data/biometrics/utils/GetBiometricsTrackedEntityAttribute.kt
@@ -1,0 +1,8 @@
+package org.dhis2.data.biometrics.utils
+
+import org.hisp.dhis.android.core.D2
+
+fun getBiometricsTrackedEntityAttribute(d2: D2): String? {
+    return d2.trackedEntityModule().trackedEntityAttributes()
+        .byDisplayFormName().eq("Biometrics").blockingGet().firstOrNull()?.uid()
+}
