+++ app/src/main/java/org/dhis2/data/biometrics/utils/GetTrackedEntityAttributeValueByAttribute.kt
@@ -1,0 +1,11 @@
+package org.dhis2.data.biometrics.utils
+
+import org.hisp.dhis.android.core.trackedentity.TrackedEntityAttributeValue
+
+fun getTrackedEntityAttributeValueByAttribute(
+    attribute: String,
+    attributeValues: List<TrackedEntityAttributeValue>
+): TrackedEntityAttributeValue? {
+    return attributeValues
+        .find { attributeValue -> attributeValue.trackedEntityAttribute() == attribute }
+}
