+++ app/src/main/java/org/dhis2/usescases/teiDashboard/DashboardRepositoryImpl.java
@@ -1,8 +1,6 @@
 package org.dhis2.usescases.teiDashboard;
 
-import static org.dhis2.commons.date.DateUtils.DATE_FORMAT_EXPRESSION;
-import static org.dhis2.data.biometrics.AddParentBiometricsAttributeValueIfRequiredKt.addParentBiometricsAttributeValueIfRequired;
-import static org.dhis2.usescases.biometrics.BiometricConstantsKt.BIOMETRICS_ENABLED;
+import static org.dhis2.data.biometrics.utils.AddParentBiometricsAttributeValueIfRequiredKt.addParentBiometricsAttributeValueIfRequired;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -12,9 +10,6 @@ import org.dhis2.commons.data.tuples.Pair;
 import org.dhis2.commons.data.tuples.Trio;
 import org.dhis2.commons.prefs.BasicPreferenceProvider;
 import org.dhis2.commons.resources.ResourceManager;
-import org.dhis2.data.biometrics.GetBiometricsParentChildConfigKt;
-import org.dhis2.usescases.biometrics.entities.BiometricsParentChildConfig;
-import org.dhis2.usescases.biometrics.entities.DateOfBirthAttributeByProgram;
 import org.dhis2.utils.AuthorityException;
 import org.dhis2.utils.DateUtils;
 import org.dhis2.utils.ValueUtils;
@@ -40,25 +35,15 @@ import org.hisp.dhis.android.core.program.ProgramIndicator;
 import org.hisp.dhis.android.core.program.ProgramRuleActionType;
 import org.hisp.dhis.android.core.program.ProgramStage;
 import org.hisp.dhis.android.core.program.ProgramTrackedEntityAttribute;
-import org.hisp.dhis.android.core.relationship.Relationship;
-import org.hisp.dhis.android.core.relationship.RelationshipItem;
-import org.hisp.dhis.android.core.relationship.RelationshipItemTrackedEntityInstance;
 import org.hisp.dhis.android.core.relationship.RelationshipType;
 import org.hisp.dhis.android.core.systeminfo.SystemInfo;
 import org.hisp.dhis.android.core.trackedentity.TrackedEntityAttribute;
 import org.hisp.dhis.android.core.trackedentity.TrackedEntityAttributeValue;
 import org.hisp.dhis.android.core.trackedentity.TrackedEntityInstance;
 import org.hisp.dhis.android.core.trackedentity.TrackedEntityType;
-import org.hisp.dhis.android.core.trackedentity.TrackedEntityTypeAttribute;
-import org.joda.time.DateTime;
-import org.joda.time.Months;
-import org.joda.time.format.DateTimeFormat;
-import org.joda.time.format.DateTimeFormatter;
 
 import java.util.ArrayList;
 import java.util.List;
-import java.util.Optional;
-import java.util.stream.Collectors;
 
 import dhis2.org.analytics.charts.Charts;
 import io.reactivex.Flowable;
